Relation change, user role
Implemented user role ( manager and employee ). Manager can manage the account - transfer and add money. User Employees can be created only by the manager.  When a new User is created it is set to a manager role, also in the same time new account is created and they are merged. After, when User manager creates employee, the same account is set to the new employee as well. In shor - there is account per manager, and unlimited employees can also be created by the managerand use the same account